agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn",
    "pydantic-ai-slim[ag-ui]",
    "python-dotenv",
    "logfire>=4.10.0",
    "github-copilot-sdk",
]

[tool.uv.sources]
pydantic-ai-slim = { path = "../../pydantic-ai/pydantic_ai_slim", editable = true }

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
